+++ src/TCWeb/Pages/Tax/Company/LossesCarriedForward.cshtml.cs
@@ -11,14 +11,13 @@ using TradeControl.Web.Authorization;
 using TradeControl.Web.Data;
 using TradeControl.Web.Models;
 
-
 namespace TradeControl.Web.Pages.Tax.Company
 {
-    public class StatementModel : DI_BasePageModel
+    public class LossesCarriedForwardModel : DI_BasePageModel
     {
-        public StatementModel(NodeContext context) : base(context) { }
+        public LossesCarriedForwardModel(NodeContext context) : base(context) { }
 
-        public IList<Cash_vwTaxCorpStatement> Cash_CorpStatement { get; set; }
+        public IList<Cash_vwTaxLossesCarriedForward> Cash_LossesCarriedForward { get; set; }
 
         public async Task OnGetAsync()
         {
@@ -26,7 +25,7 @@ namespace TradeControl.Web.Pages.Tax.Company
             {
                 await SetViewData();
 
-                Cash_CorpStatement = await NodeContext.Cash_TaxCorpStatement.OrderBy(t => t.StartOn).ToListAsync();
+                Cash_LossesCarriedForward = await NodeContext.Cash_TaxLossesCarriedForward.OrderBy(t => t.StartOn).ToListAsync();
             }
             catch (Exception e)
             {