+++ changelog.md
@@ -2,3 +2,10 @@
 
 Changes to 1.1.0, released 1 July 2021.
 
+## 1.1.1
+
+[Sql node version 3.34.7](https://github.com/tradecontrol/sqlnode/releases)
+
+- Losses carried forward on Company Tax
+- Corporation Tax totals on P&L
+- Direct invoice deletion